fix : 전체 Controller 리턴 타입 수정 / 유효성 검사 없는 API 추가

+++ src/main/java/com/supercoding/first/projectBE/controller/PostController.java
@@ -29,15 +29,20 @@ public class PostController {
 
     // 게시물 전체 조회
     @GetMapping("/posts")
-    public Map<String,List<PostResponse>> getAllPosts(){ // API 조회 예시에 맞게 맵에 리스트를 담아 리턴 하도록 수정
+    public ResponseEntity<Map<String,List<PostResponse>>> getAllPosts(@RequestHeader("Authorization") String token){ // API 조회 예시에 맞게 맵에 리스트를 담아 리턴 하도록 수정
+        String accessToken = tokenProvider.getAccessToken(token);
+
+        if(!tokenProvider.validToken(accessToken)){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
         List<PostResponse> posts = postService.getAllPosts();
         Map postsMap = new HashMap();
         postsMap.put("posts",posts);
-        return postsMap;
+        return ResponseEntity.ok().body(postsMap);
     }
     // 게시물 추가 API
     @PostMapping("/posts")
-    public ResponseEntity<String> createPost(@RequestBody PostRequest requestDto,@RequestHeader("Authorization") String token) {
+    public ResponseEntity<Map> createPost(@RequestBody PostRequest requestDto,@RequestHeader("Authorization") String token) {
         // 토큰 값에서 'Bearer ' 부분을 제거
         String jwtToken = token.substring(7);
 
@@ -47,13 +52,15 @@ public class PostController {
         }
         String email = tokenProvider.getEmail(jwtToken);
         postService.createPost(requestDto,email);
-        return new ResponseEntity<>("게시물이 성공적으로 작성 되었습니다.", HttpStatus.CREATED);
+        Map map = new HashMap();
+        map.put("message","게시물이 성공적으로 작성 되었습니다.");
+        return ResponseEntity.ok().body(map);
     }
 
 
     // 게시물 상세 조회 API
     @GetMapping("/posts/{post_id}")
-    public ResponseEntity<PostResponse> getPostById(@PathVariable Long post_id) {
+    public ResponseEntity<PostResponse> getPostById(@PathVariable Long post_id,@RequestHeader("Authorization") String token) {
         Post post = postService.getPostById(post_id);
         if (post == null) {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
@@ -65,7 +72,8 @@ public class PostController {
     @PutMapping("/posts/{post_id}")
     public ResponseEntity<PostResponse> updatePost(
             @PathVariable Long post_id,
-            @RequestBody PostRequest postRequest
+            @RequestBody PostRequest postRequest,
+            @RequestHeader("Authorization") String token
     ) throws PostNotFoundException {
         Post updatedPost = postService.updatePost(post_id, postRequest);
         if (updatedPost == null) {
@@ -76,18 +84,32 @@ public class PostController {
 
     // 게시물 삭제 API
     @DeleteMapping("/posts/{post_id}")
-    public ResponseEntity<Void> deletePost(@PathVariable Long post_id) {
+    public ResponseEntity<Map> deletePost(@PathVariable Long post_id,@RequestHeader("Authorization") String token) {
+        String accessToken = tokenProvider.getAccessToken(token);
+
+        if(!tokenProvider.validToken(accessToken)){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
         boolean deleted = postService.deletePost(post_id);
         if (!deleted) {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
-        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        Map map = new HashMap();
+        map.put("message","게시물이 삭제 되었습니다.");
+        return ResponseEntity.ok().body(map);
     }
 
     // 게시물 검색 : 작성자 이메일 API
     @GetMapping("/posts/search")
-    public List<PostResponse> getPostsByEmail(@RequestParam(name="author_email") String email) {
-        return postService.getPostsByUserEmail(email);
+    public ResponseEntity<List<PostResponse>> getPostsByEmail(@RequestParam(name="author_email") String email, @RequestHeader("Authorization") String token) {
+        String accessToken = tokenProvider.getAccessToken(token);
+
+        if(!tokenProvider.validToken(accessToken)){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+        List<PostResponse> list = postService.getPostsByUserEmail(email);
+        return ResponseEntity.ok().body(list);
     }
 
 }
